Removing floats from gaze.

+++ lib/local/GazeAnalyser/include/GazeEstimation.h
@@ -46,10 +46,10 @@ namespace GazeAnalysis
 	void DrawGaze(cv::Mat img, const LandmarkDetector::CLNF& clnf_model, cv::Point3f gazeVecAxisLeft, cv::Point3f gazeVecAxisRight, float fx, float fy, float cx, float cy);
 
 	// Getting the gaze angle in radians with respect to head pose (need to call EstimateGaze first)
-	cv::Vec2d GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& gaze_vector_2, cv::Vec6d head_pose);
+	cv::Vec2f GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& gaze_vector_2, cv::Vec6d head_pose);
 
 	// Some utilities
-	cv::Point3f GetPupilPosition(cv::Mat_<double> eyeLdmks3d);
+	cv::Point3f GetPupilPosition(cv::Mat_<float> eyeLdmks3d);
 
 }
 #endif
